Create maps for the variables
